backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-multi-agent-demo"
version = "0.1.0"
description = "Multi-tool MCP agent demo with GitHub, Web Search, and Filesystem servers"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.7",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "openai>=1.50",
    "PyGithub>=2.4",
    "tavily-python>=0.5",
    "python-dotenv>=1.0",
    "websockets>=13.0",
]

[project.scripts]
start = "backend.run:main"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
